317 - remove old TopNav and close popover when window gets smaller than 992
