+++ src/components/Layout/index.tsx
@@ -1,6 +1,6 @@
 import React, {FC, ReactNode} from 'react';
 
-import {TopNav2} from 'components';
+import {TopNav} from 'components';
 import './Layout.scss';
 
 interface ComponentProps {
@@ -11,7 +11,7 @@ const Layout: FC<ComponentProps> = ({children}) => {
   return (
     <div className="Layout">
       <div className="Layout__top-nav-wrapper">
-        <TopNav2 className="Layout__TopNav" />
+        <TopNav className="Layout__TopNav" />
       </div>
       <div className="Layout__content">{children}</div>
     </div>
