--- src/components/TopNav/TopNavDropdownMenuItem/TopNavDropdownMenuItem.scss
@@ -1,16 +1,0 @@
-.TopNavDropdownMenuItem,
-.TopNavDropdownMenuItem:visited {
-  color: var(--color-sail-gray-400);
-  display: block;
-  font-weight: 700;
-  padding: 6px 0;
-  transition: color 0.1s ease;
-
-  &:hover {
-    color: var(--color-primary);
-  }
-
-  &.active {
-    color: var(--color-primary);
-  }
-}
