--- src/components/TopNav/TopNavMenuItem/TopNavMenuItem.scss
@@ -1,20 +1,0 @@
-.TopNavMenuItem,
-.TopNavMenuItem:visited {
-  color: var(--color-sail-gray-500);
-  font-weight: 700;
-  margin-right: 36px;
-  transition: color 0.1s ease;
-
-  &:hover {
-    color: var(--color-primary);
-  }
-
-  &.active {
-    color: var(--color-primary);
-  }
-
-  @media (max-width: 992px) {
-    display: none;
-    margin: 0;
-  }
-}
