--- src/components/TopNav2/TopNavLogo/TopNavLogo.scss
@@ -1,17 +1,0 @@
-.TopNavLogo {
-  align-items: center;
-  background-color: transparent;
-  border-right: none;
-  display: flex;
-  margin-left: 24px;
-  padding: 8px 0;
-
-  &__img {
-    height: 24px;
-    margin-right: 6px;
-  }
-
-  @media (max-width: 992px) {
-    margin-left: 12px;
-  }
-}
